Feat/migration postgresql migration to profiles

+++ gamelife/src/main/java/fr/sqli/formation/gamelife/dto/ProduitDto.java
@@ -2,7 +2,6 @@ package fr.sqli.formation.gamelife.dto;
 
 import fr.sqli.formation.gamelife.entity.ImageEntity;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -22,7 +21,7 @@ public class ProduitDto {
 
 	private String plateforme;
 
-	private BigDecimal prix;
+	private int prix;
 
 	private List<ImageEntity> images;
 
@@ -74,11 +73,11 @@ public class ProduitDto {
 		this.plateforme = plateforme;
 	}
 
-	public BigDecimal getPrix() {
+	public int getPrix() {
 		return this.prix;
 	}
 
-	public void setPrix(BigDecimal prix) {
+	public void setPrix(int prix) {
 		this.prix = prix;
 	}
